requirements.txt
Babel==2.2.0
Beaker==1.8.0
Django==1.9.4
Django-Accounts==0.1
FormEncode==1.3.0
Mako==1.0.4
Markdown==2.6.5
MarkupSafe==0.23
Paste==2.0.3
PasteDeploy==1.5.2
PasteScript==2.0.2
Pillow==3.1.1
Pygments==2.0.2
Pylons==1.0.1
Routes==2.2
Tempita==0.5.2
WebError==0.13
WebHelpers==1.3
WebOb==1.6.0
WebTest==2.0.20
Werkzeug==0.11.4
argparse==1.2.1
beautifulsoup4==4.4.1
decorator==4.0.9
django-admin-bootstrapped==2.5.0
django-allauth==0.24.1
django-appconf==1.0.1
django-audit==1.0b3
django-autoslug==1.7.2
django-bootstrap-calendar==0.1.0
django-bootstrap-form==3.2
django-bootstrap-toolkit==2.15.0
django-bootstrap3==7.0.0
django-crispy-forms==1.6.0
django-email-notification==1.3
django-extensions==1.6.1
django-forms-bootstrap==3.0.1
django-geojson==2.9.0
django-grappelli==2.7.1
django-leaflet==0.18.0
django-location-field==1.6.1
django-markdown==0.8.4
django-phonenumber-field==1.0.0
django-registration==1.0
django-registration-redux==1.1
django-suit==0.2.18
djangorestframework==3.3.3
easy-thumbnails==2.3
funcsigs==0.4
icalendar==3.9.2
ipython==3.1.0
jsmin==2.2.1
nose==1.3.7
oauthlib==1.0.3
phonenumbers==7.2.7
pinax-theme-bootstrap==5.7.2
psycopg2==2.6.1
pyOpenSSL==0.15.1
pymongo==3.1.1
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2016.1
reportlab==3.3.0
repoze.lru==0.6
requests==2.9.0
requests-oauthlib==0.6.0
simplejson==3.8.2
six==1.10.0
vectorformats==0.1
waitress==0.8.10
wsgiref==0.1.2
yolk==0.4.3
